Fix the Darwin/wasm64 build and drop stray test artifacts
`std::min` could not deduce its template argument in `FunctionArraySlice`: on
Darwin and `wasm64` `size_t` and `UInt64` are distinct types (`unsigned long`
vs `unsigned long long`), so `std::min(dimension, requested_length - clipped_prefix)`
had no matching overload. Cast the `UInt64` operand to `size_t`; it is already
known to be smaller than `requested_length` at that point.

Also remove the `03371_qbit_read_write_test_*.clickhouse` files, which are
artifacts of a local run of `03371_qbit_read_write` that were committed by
mistake.

+++ src/Functions/array/arraySlice.cpp
@@ -271,7 +271,7 @@ private:
                     /// `getSliceFromRight` clamps an oversized negative offset to the left edge, but preserves the
                     /// original endpoint. A slice starting before the left edge must therefore lose the part before it.
                     const UInt64 clipped_prefix = abs_offset - dimension;
-                    result_length = requested_length > clipped_prefix ? std::min(dimension, requested_length - clipped_prefix) : 0;
+                    result_length = requested_length > clipped_prefix ? std::min(dimension, requested_length - static_cast<size_t>(clipped_prefix)) : 0;
                 }
                 else
                     result_length = std::min(result_length, requested_length);
